Create README.md
This code is a C# program that allows users to create and manage recipes. It includes the following features:

Entering recipe details: Users can enter the name, ingredients, and steps for a recipe.

Displaying a recipe: Users can display a recipe, including the ingredients, steps, and total time required to prepare the recipe.

Scaling a recipe: Users can scale a recipe up or down by a specified factor.

Resetting the quantities of ingredients: Users can reset the quantities of all ingredients in a recipe to their original values.

To use the program, simply compile and run it. The program will prompt you with the menu press 1 to enter the details of your recipe. Once you have entered the details, you can select option 2 to display the recipe. You can then scale the recipe or reset the quantities of the ingredients.

This console application is a simple but useful tool for creating and managing recipes. It can be used by anyone who enjoys cooking or baking.